+++ Sources/TossPayments/Model/PaymentInfo/MobilePhonePaymentInfo.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 public struct MobilePhonePaymentInfo: PaymentInfo {
-    public let amount: Int64
+    public let amount: Double
     public let orderId: String
     public let orderName: String
     
@@ -16,16 +16,16 @@ public struct MobilePhonePaymentInfo: PaymentInfo {
 
     public let customerName: String?
     public let customerEmail: String?
-    public let taxFreeAmount: Int64?
+    public let taxFreeAmount: Double?
     
     public init(
-        amount: Int64,
+        amount: Double,
         orderId: String,
         orderName: String,
         mobileCarrier: [CarrierCode],
         customerName: String?,
         customerEmail: String?,
-        taxFreeAmount: Int64
+        taxFreeAmount: Double
     ) {
         self.amount = amount
         self.orderId = orderId
